requirements.txt
asgiref==3.4.1
Django==3.2.8
django-active-link==0.1.8
django-cleanup==6.0.0
django-modeltranslation==0.17.3
django-phonenumber-field==6.0.0
phonenumbers==8.12.42
Pillow==8.4.0
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
